+++ axelor-base/src/main/java/com/axelor/apps/base/db/repo/TeamTaskBaseRepository.java
@@ -17,7 +17,7 @@
  */
 package com.axelor.apps.base.db.repo;
 
-import com.axelor.apps.base.db.Frequency;
+import com.axelor.apps.base.db.Repetition;
 import com.axelor.apps.base.service.TeamTaskService;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
@@ -33,7 +33,7 @@ public class TeamTaskBaseRepository extends TeamTaskRepository {
 
     if (teamTask.getDoApplyToAllNextTasks()
         && teamTask.getNextTeamTask() != null
-        && teamTask.getHasDateOrFrequencyChanged()) {
+        && teamTask.getHasDateOrRepetitionChanged()) {
       // remove next tasks
       teamTaskService.removeNextTasks(teamTask);
 
@@ -41,18 +41,18 @@ public class TeamTaskBaseRepository extends TeamTaskRepository {
       teamTask.setIsFirst(true);
     }
 
-    Frequency frequency = teamTask.getFrequency();
-    if (frequency != null && teamTask.getIsFirst() && teamTask.getNextTeamTask() == null) {
+    Repetition repetition = teamTask.getRepetition();
+    if (repetition != null && teamTask.getIsFirst() && teamTask.getNextTeamTask() == null) {
       if (teamTask.getTaskDate() != null) {
-        if (frequency.getEndDate().isBefore(teamTask.getTaskDate())) {
+        if (teamTask.getRepetitionEndDate().isBefore(teamTask.getTaskDate())) {
           throw new PersistenceException(
-              I18n.get("Frequency end date cannot be before task date."));
+              I18n.get("Repetition end date cannot be before task date."));
         }
       } else {
         throw new PersistenceException(I18n.get("Please fill in task date."));
       }
 
-      teamTaskService.generateTasks(teamTask, frequency);
+      teamTaskService.generateTasks(teamTask, repetition);
     }
 
     if (teamTask.getDoApplyToAllNextTasks()) {
@@ -60,7 +60,7 @@ public class TeamTaskBaseRepository extends TeamTaskRepository {
     }
 
     teamTask.setDoApplyToAllNextTasks(false);
-    teamTask.setHasDateOrFrequencyChanged(false);
+    teamTask.setHasDateOrRepetitionChanged(false);
 
     return super.save(teamTask);
   }