+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased 5.2.0]
 ## Features
 - Employee: added DPAE
-- TeamTask: added Frequency
+- TeamTask: added Repetition
 - Employee: added wizard when creating new employee
 - LogisticalForm: if config enabled, send an email on first save
 - SaleOrder: added boolean to invoice via generated task